+++ src/main/java/com/scanii/ScaniiClient.java
@@ -14,7 +14,7 @@ import java.util.concurrent.TimeUnit;
 public interface ScaniiClient {
 
   /**
-   * Submits a file to be processed @see <a href="http://docs.scanii.com/v2.1/resources.html#files">http://docs.scanii.com/v2.1/resources.html#files</a>
+   * Submits a file to be processed @see <a href="https://docs.scanii.com/v2.2/resources.html#files">https://docs.scanii.com/v2.2/resources.html#files</a>
    *
    * @param content  path to the file to be processed
    * @param metadata optional metadata to be added to this file
@@ -23,7 +23,7 @@ public interface ScaniiClient {
   ScaniiProcessingResult process(Path content, Map<String, String> metadata);
 
   /**
-   * Submits a file stream to be processed @see <a href="http://docs.scanii.com/v2.1/resources.html#files">http://docs.scanii.com/v2.1/resources.html#files</a>
+   * Submits a file stream to be processed @see <a href="https://docs.scanii.com/v2.2/resources.html#files">https://docs.scanii.com/v2.2/resources.html#files</a>
    *
    * @param content  stream of the file to be processed
    * @param metadata optional metadata to be added to this file
@@ -32,7 +32,7 @@ public interface ScaniiClient {
   ScaniiProcessingResult process(InputStream content, Map<String, String> metadata);
 
   /**
-   * Submits a file to be processed @see <a href="http://docs.scanii.com/v2.1/resources.html#files">http://docs.scanii.com/v2.1/resources.html#files</a> with optional callback.
+   * Submits a file to be processed @see <a href="https://docs.scanii.com/v2.2/resources.html#files">https://docs.scanii.com/v2.2/resources.html#files</a> with optional callback.
    *
    * @param content  path to the file to be processed
    * @param callback location (URL) to be notified and receive the result
@@ -43,7 +43,7 @@ public interface ScaniiClient {
 
 
   /**
-   * Submits a file stream to be processed @see <a href="http://docs.scanii.com/v2.1/resources.html#files">http://docs.scanii.com/v2.1/resources.html#files</a> with optional callback.
+   * Submits a file stream to be processed @see <a href="https://docs.scanii.com/v2.2/resources.html#files">https://docs.scanii.com/v2.2/resources.html#files</a> with optional callback.
    *
    * @param content  file content to be processed
    * @param callback location (URL) to be notified and receive the result
@@ -53,7 +53,7 @@ public interface ScaniiClient {
   ScaniiProcessingResult process(InputStream content, String callback, Map<String, String> metadata);
 
   /**
-   * Submits a file to be processed @see <a href="http://docs.scanii.com/v2.1/resources.html#files">http://docs.scanii.com/v2.1/resources.html#files</a>
+   * Submits a file to be processed @see <a href="https://docs.scanii.com/v2.2/resources.html#files">https://docs.scanii.com/v2.2/resources.html#files</a>
    *
    * @param content path to the file to be processed
    * @return scanii result {@link ScaniiProcessingResult}
@@ -61,7 +61,7 @@ public interface ScaniiClient {
   ScaniiProcessingResult process(Path content);
 
   /**
-   * Submits a file stream to be processed @see <a href="http://docs.scanii.com/v2.1/resources.html#files">http://docs.scanii.com/v2.1/resources.html#files</a>
+   * Submits a file stream to be processed @see <a href="https://docs.scanii.com/v2.2/resources.html#files">https://docs.scanii.com/v2.2/resources.html#files</a>
    *
    * @param content stream of the file to be processed
    * @return scanii result {@link ScaniiProcessingResult}
@@ -69,7 +69,7 @@ public interface ScaniiClient {
   ScaniiProcessingResult process(InputStream content);
 
   /**
-   * Submits a file to be processed asynchronously @see <a href="http://docs.scanii.com/v2.1/resources.html#files">http://docs.scanii.com/v2.1/resources.html#files</a>
+   * Submits a file to be processed asynchronously @see <a href="https://docs.scanii.com/v2.2/resources.html#files">https://docs.scanii.com/v2.2/resources.html#files</a>
    *
    * @param content  path to the file to be processed
    * @param metadata optional metadata to be added to this file
@@ -78,7 +78,7 @@ public interface ScaniiClient {
   ScaniiPendingResult processAsync(Path content, Map<String, String> metadata);
 
   /**
-   * Submits a file stream to be processed asynchronously @see <a href="http://docs.scanii.com/v2.1/resources.html#files">http://docs.scanii.com/v2.1/resources.html#files</a>
+   * Submits a file stream to be processed asynchronously @see <a href="https://docs.scanii.com/v2.2/resources.html#files">https://docs.scanii.com/v2.2/resources.html#files</a>
    *
    * @param content  stream of the file to be processed
    * @param metadata optional metadata to be added to this file
@@ -87,7 +87,7 @@ public interface ScaniiClient {
   ScaniiPendingResult processAsync(InputStream content, Map<String, String> metadata);
 
   /**
-   * Submits a file to be processed asynchronously @see <a href="http://docs.scanii.com/v2.1/resources.html#files">http://docs.scanii.com/v2.1/resources.html#files</a>
+   * Submits a file to be processed asynchronously @see <a href="https://docs.scanii.com/v2.2/resources.html#files">https://docs.scanii.com/v2.2/resources.html#files</a>
    *
    * @param content  path to the file to be processed
    * @param callback location (URL) to be notified and receive the result
@@ -97,7 +97,7 @@ public interface ScaniiClient {
   ScaniiPendingResult processAsync(Path content, String callback, Map<String, String> metadata);
 
   /**
-   * Submits a file stream to be processed @see <a href="http://docs.scanii.com/v2.1/resources.html#files">http://docs.scanii.com/v2.1/resources.html#files</a> with optional callback.
+   * Submits a file stream to be processed @see <a href="https://docs.scanii.com/v2.2/resources.html#files">https://docs.scanii.com/v2.2/resources.html#files</a> with optional callback.
    *
    * @param content  file content to be processed
    * @param callback location (URL) to be notified and receive the result
@@ -107,7 +107,7 @@ public interface ScaniiClient {
   ScaniiPendingResult processAsync(InputStream content, String callback, Map<String, String> metadata);
 
   /**
-   * Submits a file to be processed asynchronously @see <a href="http://docs.scanii.com/v2.1/resources.html#files">http://docs.scanii.com/v2.1/resources.html#files</a>
+   * Submits a file to be processed asynchronously @see <a href="https://docs.scanii.com/v2.2/resources.html#files">https://docs.scanii.com/v2.2/resources.html#files</a>
    *
    * @param content path to the file to be processed
    * @return processing result {@link ScaniiPendingResult}
@@ -115,7 +115,7 @@ public interface ScaniiClient {
   ScaniiPendingResult processAsync(Path content);
 
   /**
-   * Submits a file content stream to be processed asynchronously @see <a href="http://docs.scanii.com/v2.1/resources.html#files">http://docs.scanii.com/v2.1/resources.html#files</a>
+   * Submits a file content stream to be processed asynchronously @see <a href="https://docs.scanii.com/v2.2/resources.html#files">https://docs.scanii.com/v2.2/resources.html#files</a>
    *
    * @param content stream of the file to be processed
    * @return processing result {@link ScaniiPendingResult}
@@ -123,7 +123,7 @@ public interface ScaniiClient {
   ScaniiPendingResult processAsync(InputStream content);
 
   /**
-   * Fetches the results of a previously processed file @see <a href="http://docs.scanii.com/v2.1/resources.html#files">http://docs.scanii.com/v2.1/resources.html#files</a>
+   * Fetches the results of a previously processed file @see <a href="https://docs.scanii.com/v2.2/resources.html#files">https://docs.scanii.com/v2.2/resources.html#files</a>
    *
    * @param id id of the content/file to be retrieved
    * @return optional  {@link ScaniiProcessingResult}
@@ -131,7 +131,7 @@ public interface ScaniiClient {
   Optional<ScaniiProcessingResult> retrieve(String id);
 
   /**
-   * Makes a fetch call to scanii @see <a href="http://docs.scanii.com/v2.1/resources.html#files">http://docs.scanii.com/v2.1/resources.html#files</a>
+   * Makes a fetch call to scanii @see <a href="https://docs.scanii.com/v2.2/resources.html#files">https://docs.scanii.com/v2.2/resources.html#files</a>
    *
    * @param location location (URL) of the content to be processed
    * @return scanii result {@link ScaniiPendingResult}
@@ -139,7 +139,7 @@ public interface ScaniiClient {
   ScaniiPendingResult fetch(String location);
 
   /**
-   * Makes a fetch call to scanii @see <a href="http://docs.scanii.com/v2.1/resources.html#files">http://docs.scanii.com/v2.1/resources.html#files</a>
+   * Makes a fetch call to scanii @see <a href="https://docs.scanii.com/v2.2/resources.html#files">https://docs.scanii.com/v2.2/resources.html#files</a>
    *
    * @param location location (URL) of the content to be processed
    * @param callback location (URL) to be notified and receive the result
@@ -148,7 +148,7 @@ public interface ScaniiClient {
   ScaniiPendingResult fetch(String location, String callback);
 
   /**
-   * Makes a fetch call to scanii @see <a href="http://docs.scanii.com/v2.1/resources.html#files">http://docs.scanii.com/v2.1/resources.html#files</a>
+   * Makes a fetch call to scanii @see <a href="https://docs.scanii.com/v2.2/resources.html#files">https://docs.scanii.com/v2.2/resources.html#files</a>
    *
    * @param location location (URL) of the content to be processed
    * @param callback location (URL) to be notified and receive the result
@@ -158,14 +158,14 @@ public interface ScaniiClient {
   ScaniiPendingResult fetch(String location, String callback, Map<String, String> metadata);
 
   /**
-   * Pings the scanii service using the credentials provided @see <a href="http://docs.scanii.com/v2.1/resources.html#ping">http://docs.scanii.com/v2.1/resources.html#ping</a>
+   * Pings the scanii service using the credentials provided @see <a href="https://docs.scanii.com/v2.2/resources.html#ping">https://docs.scanii.com/v2.2/resources.html#ping</a>
    *
    * @return true if we saw a pong back from scanii, false otherwise
    */
   boolean ping();
 
   /**
-   * Creates a new temporary authentication token @see <a href="http://docs.scanii.com/v2.1/resources.html#auth-tokens">http://docs.scanii.com/v2.1/resources.html#auth-tokens</a>
+   * Creates a new temporary authentication token @see <a href="https://docs.scanii.com/v2.2/resources.html#auth-tokens">https://docs.scanii.com/v2.2/resources.html#auth-tokens</a>
    *
    * @param timeout     how long the token should be valid for"
    * @param timeoutUnit unit used to calculate the timeout
@@ -189,7 +189,7 @@ public interface ScaniiClient {
   ScaniiAuthToken retrieveAuthToken(String id);
 
   /**
-   * Retrieves account information @see <a href="https://docs.scanii.com/v2.1/resources.html#account">https://docs.scanii.com/v2.1/resources.html#account</a>
+   * Retrieves account information @see <a href="https://docs.scanii.com/v2.2/resources.html#account">https://docs.scanii.com/v2.2/resources.html#account</a>
    *
    * @return your account information
    */